Move auth code into login crate
- Move the auth implementation and token data into `codex-login` instead of a standalone auth crate.
- Keep `codex-core` re-exporting that surface from `codex-login` for existing callers.
