--- codex-rs/auth/BUILD.bazel
@@ -1,7 +1,0 @@
-load("//:defs.bzl", "codex_rust_crate")
-
-codex_rust_crate(
-    name = "auth",
-    crate_name = "codex_auth",
-    crate_srcs = glob(["*.rs"]),
-)
