+++ codex-rs/login/src/auth/auth.rs
@@ -17,17 +17,17 @@ use codex_app_server_protocol::AuthMode as ApiAuthMode;
 use codex_otel::TelemetryAuthMode;
 use codex_protocol::config_types::ForcedLoginMethod;
 
-use crate::error::RefreshTokenFailedError;
-use crate::error::RefreshTokenFailedReason;
-pub use crate::storage::AuthCredentialsStoreMode;
-pub use crate::storage::AuthDotJson;
-use crate::storage::AuthStorageBackend;
-use crate::storage::create_auth_storage;
+use crate::auth::error::RefreshTokenFailedError;
+use crate::auth::error::RefreshTokenFailedReason;
+pub use crate::auth::storage::AuthCredentialsStoreMode;
+pub use crate::auth::storage::AuthDotJson;
+use crate::auth::storage::AuthStorageBackend;
+use crate::auth::storage::create_auth_storage;
+use crate::auth::util::try_parse_error_message;
 use crate::token_data::KnownPlan as InternalKnownPlan;
 use crate::token_data::PlanType as InternalPlanType;
 use crate::token_data::TokenData;
 use crate::token_data::parse_chatgpt_jwt_claims;
-use crate::util::try_parse_error_message;
 use codex_client::CodexHttpClient;
 use codex_protocol::account::PlanType as AccountPlanType;
 use serde_json::Value;
