+++ codex-rs/login/src/auth/error.rs
