+++ codex-rs/login/src/auth/mod.rs
@@ -1,6 +1,5 @@
 pub mod error;
 mod storage;
-pub mod token_data;
 mod util;
 
 mod auth;
