+++ codex-rs/login/src/auth/storage_tests.rs
