+++ codex-rs/login/src/auth/util.rs
